refactor: migrate toast system from src/shared/toast to src/modules/toast with clean architecture layers (domain, application, infrastructure, ui); update all imports to use ~/ absolute paths; move and update all related tests
