+++ src/modules/diet/recipe/application/recipe.ts
@@ -1,7 +1,7 @@
 import { createSupabaseRecipeRepository } from '~/modules/diet/recipe/infrastructure/supabaseRecipeRepository'
 import { type User } from '~/modules/user/domain/user'
 import { type Recipe, type NewRecipe } from '../domain/recipe'
-import { showPromise } from '~/shared/toast/toastManager'
+import { showPromise } from '~/modules/toast/application/toastManager'
 import { handleApiError } from '~/shared/error/errorHandler'
 
 const recipeRepository = createSupabaseRecipeRepository()
