+++ src/modules/toast/infrastructure/clipboardErrorUtils.ts
@@ -2,10 +2,10 @@
  * Utilities for formatting and copying error details to clipboard.
  * Shared between ExpandableErrorToast and ErrorDetailModal.
  */
-import { ToastError } from './toastConfig'
+import { ToastError } from '~/modules/toast/domain/toastTypes'
 import { handleApiError } from '~/shared/error/errorHandler'
 import { useClipboard } from '~/sections/common/hooks/useClipboard'
-import { TOAST_MESSAGES } from './toastMessages'
+import { TOAST_MESSAGES } from '~/modules/toast/domain/toastMessages'
 
 // Explicit type for context
 export type ClipboardErrorContext = { component: string; operation: string }
