+++ src/modules/toast/infrastructure/toastSettings.ts
