+++ src/routes/settings.tsx
@@ -1,5 +1,5 @@
 import { For, createSignal } from 'solid-js'
-import { showSuccess } from '~/shared/toast/toastManager'
+import { showSuccess } from '~/modules/toast/application/toastManager'
 import { CARD_BACKGROUND_COLOR, CARD_STYLE } from '~/modules/theme/constants'
 import { BottomNavigation } from '~/sections/common/components/BottomNavigation'
 import { Providers } from '~/sections/common/context/Providers'
