+++ src/sections/common/context/Providers.tsx
@@ -1,7 +1,7 @@
 import { ConfirmModalProvider } from './ConfirmModalContext'
 import { ConfirmModal } from '../components/ConfirmModal'
 import { DarkToaster } from '../components/DarkToaster'
-import { GlobalModalContainer } from '~/shared/toast/GlobalModalContainer'
+import { GlobalModalContainer } from '~/modules/toast/ui/GlobalModalContainer'
 import { type JSXElement } from 'solid-js'
 
 export function Providers(props: { children: JSXElement }) {
