+++ src/sections/day-diet/components/CopyLastDayButton.tsx
@@ -9,7 +9,7 @@ import {
   updateDayDiet,
 } from '~/modules/diet/day-diet/application/dayDiet'
 import { Show, createEffect, createSignal, type Accessor } from 'solid-js'
-import { showError } from '~/shared/toast/toastManager'
+import { showError } from '~/modules/toast/application/toastManager'
 
 export function CopyLastDayButton(props: {
   dayDiet: Accessor<DayDiet | undefined> // TODO:   Rename all 'day' to 'dayDiet' on entire project.
