+++ src/sections/item-group/components/ItemGroupEditModal.tsx
@@ -72,7 +72,7 @@ import { getMacroTargetForDay } from '~/modules/diet/macro-target/application/ma
 import { stringToDate } from '~/shared/utils/date'
 import { type MacroNutrients } from '~/modules/diet/macro-nutrients/domain/macroNutrients'
 import { isOverflow } from '~/legacy/utils/macroOverflow'
-import { showError } from '~/shared/toast/toastManager'
+import { showError } from '~/modules/toast/application/toastManager'
 import {
   fetchRecipeById,
   insertRecipe,
