+++ src/sections/search/components/TemplateSearchModal.tsx
@@ -18,7 +18,7 @@ import { createNewRecentFood } from '~/modules/recent-food/domain/recentFood'
 import { type Template } from '~/modules/diet/template/domain/template'
 import { handleApiError } from '~/shared/error/errorHandler'
 import { type TemplateItem } from '~/modules/diet/template-item/domain/templateItem'
-import { showSuccess } from '~/shared/toast/toastManager'
+import { showSuccess } from '~/modules/toast/application/toastManager'
 
 import { type Food } from '~/modules/diet/food/domain/food'
 import { currentUser, currentUserId } from '~/modules/user/application/user'
@@ -50,7 +50,7 @@ import { getMacroTargetForDay } from '~/modules/diet/macro-target/application/ma
 import { stringToDate } from '~/shared/utils/date'
 import { isOverflowForItemGroup } from '~/legacy/utils/macroOverflow'
 import { type MacroNutrients } from '~/modules/diet/macro-nutrients/domain/macroNutrients'
-import { showError } from '~/shared/toast/toastManager'
+import { showError } from '~/modules/toast/application/toastManager'
 import { TemplateSearchBar } from './TemplateSearchBar'
 import { TemplateSearchResults } from './TemplateSearchResults'
 import { BarCodeButton } from '~/sections/common/components/BarCodeButton'
